+++ src/gauge/csv_printer.hpp
@@ -14,19 +14,9 @@ namespace gauge
         /// @param default_filename The default name of the outputted file
         csv_printer(const std::string& default_filename = "out.csv");
 
-    public: // From printer
-
-        /// @see printer::set_options(po::variables_map&);
-        void set_options(po::variables_map& options);
-
     public: // From file_printer
 
         /// @see file_printer::print_to_stream(std::ostream &s)
         void print_to_stream(std::ostream &s);
-
-    private:
-
-        /// Store the value separator
-        std::string m_value_seperator;
     };
 }